requirements.txt
torch
torchvision
pytorch-lightning
lightning
scikit-learn
tqdm
matplotlib
numpy
scipy
scikit-image
wandb
pyyaml
pillow
fire
py4DSTEM